move TokenHasScope to correct app
